+++ experimental/schemabuilder/example/query.go
