+++ experimental/schemabuilder/example/schema.go
