+++ experimental/schemabuilder/testdata/v0alpha1/query.examples.json
