+++ experimental/schemabuilder/testdata/v0alpha1/routes.yaml
