+++ experimental/schemabuilder/testdata/v0alpha1/settings.examples.json
