+++ experimental/schemabuilder/testdata/v0alpha1/settings.json
